reset targetDate after a form reset with no default value
Currently, when no default value is set, a form "reset" event clears the input but does not reset the label datetime. So after clicking the reset button, opening the date picker shows the last date selected before the reset. This modification aims to fully reset the calendar state after a form reset when no default value is set.